Change from Long to BigInt for id values
To better match the JSON serialisation library, and also remove the
possibility of exceeding Long value limits

+++ hypedyn-core/src/main/scala/org/narrativeandplay/hypedyn/serialisation/serialisers/JsonSerialiser.scala
@@ -24,7 +24,7 @@ object JsonSerialiser {
   }
 
   private def jValueToAstElement(jValue: JValue): AstElement = jValue match {
-    case JInt(i) => AstInteger(i.toLong)
+    case JInt(i) => AstInteger(i)
     case JDouble(f) => AstFloat(f)
     case JString(s) => AstString(s)
     case JBool(b) => AstBoolean(b)
